scripts/hyperparameter_tuning.py
import optuna
from optuna.integration import MLflowCallback
import mlflow
import yaml
import sys
sys.path.append('.')

from src.model_training import ModelTrainer
from src.data_preprocessing import DataPreprocessor

def objective(trial):
    """Optuna objective function"""
    
    # Suggest hyperparameters
    params = {
        'n_estimators': trial.suggest_int('n_estimators', 100, 500),
        'max_depth': trial.suggest_int('max_depth', 3, 10),
        'learning_rate': trial.suggest_float('learning_rate', 0.01, 0.3),
        'subsample': trial.suggest_float('subsample', 0.6, 1.0),
        'colsample_bytree': trial.suggest_float('colsample_bytree', 0.6, 1.0)
    }
    
    # Train and evaluate
    # (Add your training code here)
    
    return roc_auc_score  # Return metric to optimize

def run_tuning():
    print("🔧 Starting Hyperparameter Tuning")
    
    study = optuna.create_study(direction='maximize')
    study.optimize(objective, n_trials=100, timeout=3600)
    
    print(f"\n✅ Best parameters: {study.best_params}")
    print(f"✅ Best ROC-AUC: {study.best_value:.4f}")

if __name__ == "__main__":
    run_tuning()
