requirements.txt
# Core API
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.2
lightgbm==4.1.0
catboost==1.2.2

# Data Processing
pandas==2.1.3
numpy==1.26.2
imbalanced-learn==0.11.0

# Model Management
mlflow==2.8.1
joblib==1.3.2

# Explainability
shap==0.43.0
lime==0.2.0.1

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Hyperparameter Optimization
optuna==3.4.0

# Statistical Analysis
scipy==1.11.4
statsmodels==0.14.1

# Utilities
pyyaml==6.0.1
requests==2.31.0